+++ meteo_lt/__init__.py
@@ -1,6 +1,13 @@
 """init.py"""
 
 from .api import MeteoLtAPI
-from .models import Coordinates, Place, ForecastTimestamp, Forecast
+from .models import Coordinates, Place, ForecastTimestamp, Forecast, WeatherWarning
 
-__all__ = ["MeteoLtAPI", "Coordinates", "Place", "ForecastTimestamp", "Forecast"]
+__all__ = [
+    "MeteoLtAPI",
+    "Coordinates",
+    "Place",
+    "ForecastTimestamp",
+    "Forecast",
+    "WeatherWarning",
+]